Correctly normalize css for server side rendering
This fixes #1 to make server side rendering work correctly. Previously
there was a bug where, because the css dynamically created didn't
contain the `!` indicator, server-side rendering would not pick up the
style as a css dependency and it would not get created. This fixes it by
making sure it normalizes correctly. Also adds an ssr test.