Refactor to define virtual modules in a fetch hook
This refactor makes it so that `loader.define` is not called during
translate. The problem with loader.define being called is that it will
go all the way through instantiate and during live-reload / watch module
we do a translate to check if dependencies have changed.

This change makes it so that the virtual modules are not defined, but
rather their source is kept in a side table `loader.componentSources`
and a fetch hook returns them.

Additionally each .component's virtual dependencies are stored in
`load.metadata.virtualDeps`. The plan is for steal-tools to use that
info to determine if any of the virtual modules dependencies have
changed.

Closes #8