Define fetch hook on both loaders
In the build there are two loaders, the plugin loader (which is
`@loader` here) and the "localLoader" which is the main loader
doing the trace. We need to attach the new fetch hook to both loaders so
that the virtual modules are properly fetched and added to the
dependency graph.